Feat: Add workspace create handlers
Added actions, reducer to handle creation of Workspace using a temporary ID generated from a utility function.

The action is not yet complete and is awaiting on the backend to finalise the POST route in order for the service to be implemented and the whole creation flow can be added.